+++ src/uucp_script.ml
@@ -6,7 +6,7 @@
 
 include Uucp_script_base
 
-let compare = Pervasives.compare
+let compare = compare
 let script u = Uucp_rmap.get Uucp_script_data.script_map (Uchar.to_int u)
 let script_extensions u =
   match Uucp_rmap.get Uucp_script_data.script_extensions_map (Uchar.to_int u)