+++ support/gen_props.ml
@@ -30,7 +30,7 @@ let process
   | None -> ()
   | Some fn ->
       try
-        let oc = if fn = "-" then stdout else Pervasives.open_out fn in
+        let oc = if fn = "-" then stdout else open_out fn in
         try
           let ppf = Format.formatter_of_out_channel oc in
           pp ppf ucd;